# docs/api/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "NISQ Analyzer API",
    "contact" : {
      "name" : "GitHub Repository",
      "url" : "https://github.com/UST-QuAntiL/nisq-analyzer"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.4.0"
  },
  "servers" : [ {
    "url" : "http://localhost:5010/nisq-analyzer",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/execution-results/" : {
      "get" : {
        "tags" : [ "execution-result" ],
        "description" : "Retrieve all execution results for an Implementation",
        "operationId" : "getExecutionResults",
        "parameters" : [ {
          "name" : "implId",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExecutionResultListDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/implementations/" : {
      "get" : {
        "tags" : [ "implementation" ],
        "description" : "Retrieve implementations for an algorithm",
        "operationId" : "getImplementations",
        "parameters" : [ {
          "name" : "algoId",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImplementationListDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      },
      "post" : {
        "tags" : [ "implementation" ],
        "description" : "Create an implementation",
        "operationId" : "createImplementation",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ImplementationDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImplementationDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        }
      }
    },
    "/implementations/{implId}/output-parameters" : {
      "get" : {
        "tags" : [ "implementation" ],
        "description" : "Retrieve output parameters for an implementation",
        "operationId" : "getOutputParameters",
        "parameters" : [ {
          "name" : "implId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ParameterListDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      },
      "post" : {
        "tags" : [ "implementation" ],
        "description" : "Add output parameters to an implementation",
        "operationId" : "addOutputParameter",
        "parameters" : [ {
          "name" : "implId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ParameterDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ParameterDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/implementations/{implId}" : {
      "get" : {
        "tags" : [ "implementation" ],
        "description" : "Retrieve an implementation",
        "operationId" : "getImplementation",
        "parameters" : [ {
          "name" : "implId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImplementationDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      },
      "put" : {
        "tags" : [ "implementation" ],
        "description" : "Update an implementation",
        "operationId" : "updateImplementation",
        "parameters" : [ {
          "name" : "implId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ImplementationDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImplementationDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        }
      }
    },
    "/implementations/{implId}/input-parameters" : {
      "get" : {
        "tags" : [ "implementation" ],
        "description" : "Retrieve input parameters for an implementation",
        "operationId" : "getInputParameters",
        "parameters" : [ {
          "name" : "implId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ParameterListDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      },
      "post" : {
        "tags" : [ "implementation" ],
        "description" : "Add input parameters to an implementation",
        "operationId" : "addInputParameter",
        "parameters" : [ {
          "name" : "implId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ParameterDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ParameterDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      },
      "delete" : {
        "tags" : [ "implementation" ],
        "description" : "Remove input parameters from an implementation",
        "operationId" : "deleteInputParameters",
        "parameters" : [ {
          "name" : "implId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ImplementationDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "implementedAlgorithm" : {
            "type" : "string",
            "format" : "uuid"
          },
          "algorithmName" : {
            "type" : "string"
          },
          "language" : {
            "type" : "string"
          },
          "sdk" : {
            "type" : "string"
          },
          "fileLocation" : {
            "type" : "string",
            "format" : "url"
          },
          "inputParameters" : {
            "$ref" : "#/components/schemas/ParameterListDto"
          },
          "outputParameters" : {
            "$ref" : "#/components/schemas/ParameterListDto"
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
      "Links" : {
        "type" : "object",
        "additionalProperties" : {
          "$ref" : "#/components/schemas/Link"
        }
      },
      "ParameterDto" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "Integer", "Float", "IntegerArray", "FloatArray", "String", "Unknown" ]
          },
          "restriction" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          }
        }
      },
      "ParameterListDto" : {
        "type" : "object",
        "properties" : {
          "parameters" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ParameterDto"
            }
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
...
      "ExecutionResultDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "INITIALIZED", "RUNNING", "FAILED", "FINISHED" ]
          },
          "statusCode" : {
            "type" : "string"
          },
          "result" : {
            "type" : "string"
          },
          "shots" : {
            "type" : "integer",
            "format" : "int32"
          },
          "histogramIntersectionValue" : {
            "type" : "number",
            "format" : "double"
          },
          "resultLocation" : {
            "type" : "string",
            "format" : "uri"
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
      "ExecutionResultListDto" : {
        "type" : "object",
        "properties" : {
          "executionResultDtos" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ExecutionResultDto"
            }
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
      "ImplementationListDto" : {
        "type" : "object",
        "properties" : {
          "implementationDtos" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ImplementationDto"
            }
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
...
      "Link" : {
        "type" : "object",
        "properties" : {
          "href" : {
            "type" : "string"
          },
          "hreflang" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "deprecation" : {
            "type" : "string"
          },
          "profile" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "templated" : {
            "type" : "boolean"
          }
        }
      },
...
    }
  }
}